% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery_functions.R
\name{apis.getRest}
\alias{apis.getRest}
\title{Retrieve the description of a particular version of an api.}
\usage{
apis.getRest(api, version)
}
\arguments{
\item{api}{The name of the API}

\item{version}{The version of the API}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item 
}

Set \code{options(googleAuthR.scopes.selected = c()}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/discovery/}{Google Documentation}
}
