% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery_functions.R
\name{apis.list}
\alias{apis.list}
\title{Retrieve the list of APIs supported at this endpoint.}
\usage{
apis.list(name = NULL, preferred = NULL)
}
\arguments{
\item{name}{Only include APIs with the given name}

\item{preferred}{Return only the preferred version of an API}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item 
}

Set \code{options(googleAuthR.scopes.selected = c()}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/discovery/}{Google Documentation}
}

