% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery_objects.R
\name{DirectoryList.items.icons}
\alias{DirectoryList.items.icons}
\title{DirectoryList.items.icons Object}
\usage{
DirectoryList.items.icons(x16 = NULL, x32 = NULL)
}
\arguments{
\item{x16}{The URL of the 16x16 icon}

\item{x32}{The URL of the 32x32 icon}
}
\value{
DirectoryList.items.icons object
}
\description{
DirectoryList.items.icons Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Links to 16x16 and 32x32 icons representing the API.
}
\seealso{
Other DirectoryList functions: \code{\link{DirectoryList.items}},
  \code{\link{DirectoryList}}
}

