% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery_objects.R
\name{DirectoryList.items}
\alias{DirectoryList.items}
\title{DirectoryList.items Object}
\usage{
DirectoryList.items(DirectoryList.items.icons = NULL)
}
\arguments{
\item{DirectoryList.items.icons}{The \link{DirectoryList.items.icons} object or list of objects}
}
\value{
DirectoryList.items object
}
\description{
DirectoryList.items Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other DirectoryList functions: \code{\link{DirectoryList.items.icons}},
  \code{\link{DirectoryList}}
}
