% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery_objects.R
\name{DirectoryList}
\alias{DirectoryList}
\title{APIs Discovery Service Objects 
Provides information about other Google APIs, such as what APIs are available, the resource, and method details for each API.}
\usage{
DirectoryList(DirectoryList.items = NULL, DirectoryList.items.icons = NULL,
  discoveryVersion = NULL, items = NULL)
}
\arguments{
\item{DirectoryList.items}{The \link{DirectoryList.items} object or list of objects}

\item{DirectoryList.items.icons}{The \link{DirectoryList.items.icons} object or list of objects}

\item{discoveryVersion}{Indicate the version of the Discovery API used to generate this doc}

\item{items}{The individual directory entries}
}
\value{
DirectoryList object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-03 23:10:31
filename: /Users/mark/dev/R/autoGoogleAPI/googlediscoveryv1.auto/R/discovery_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
DirectoryList Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other DirectoryList functions: \code{\link{DirectoryList.items.icons}},
  \code{\link{DirectoryList.items}}
}

