% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery_functions.R
\docType{package}
\name{discovery_googleAuthR}
\alias{discovery_googleAuthR}
\alias{discovery_googleAuthR-package}
\title{APIs Discovery Service
Provides information about other Google APIs, such as what APIs are available, the resource, and method details for each API.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2016-09-03 23:10:31
filename: /Users/mark/dev/R/autoGoogleAPI/googlediscoveryv1.auto/R/discovery_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item 
}
}

