% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery_objects.R
\name{JsonSchema.annotations}
\alias{JsonSchema.annotations}
\title{JsonSchema.annotations Object}
\usage{
JsonSchema.annotations(required = NULL)
}
\arguments{
\item{required}{A list of methods for which this property is required on requests}
}
\value{
JsonSchema.annotations object
}
\description{
JsonSchema.annotations Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Additional information about this property.
}
\seealso{
Other JsonSchema functions: \code{\link{JsonSchema.properties}},
  \code{\link{JsonSchema.variant.map}},
  \code{\link{JsonSchema.variant}},
  \code{\link{JsonSchema}}
}
