% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery_objects.R
\name{JsonSchema.properties}
\alias{JsonSchema.properties}
\title{JsonSchema.properties Object}
\usage{
JsonSchema.properties()
}
\value{
JsonSchema.properties object
}
\description{
JsonSchema.properties Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
If this is a schema for an object, list the schema for each property of this object.
}
\seealso{
Other JsonSchema functions: \code{\link{JsonSchema.annotations}},
  \code{\link{JsonSchema.variant.map}},
  \code{\link{JsonSchema.variant}},
  \code{\link{JsonSchema}}
}

