% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery_objects.R
\name{JsonSchema}
\alias{JsonSchema}
\title{JsonSchema Object}
\usage{
JsonSchema(JsonSchema.annotations = NULL, JsonSchema.properties = NULL,
  JsonSchema.variant = NULL, JsonSchema.variant.map = NULL, X.ref = NULL,
  additionalProperties = NULL, annotations = NULL, default = NULL,
  description = NULL, enum = NULL, enumDescriptions = NULL,
  format = NULL, id = NULL, items = NULL, location = NULL,
  maximum = NULL, minimum = NULL, pattern = NULL, properties = NULL,
  readOnly = NULL, repeated = NULL, required = NULL, type = NULL,
  variant = NULL)
}
\arguments{
\item{JsonSchema.annotations}{The \link{JsonSchema.annotations} object or list of objects}

\item{JsonSchema.properties}{The \link{JsonSchema.properties} object or list of objects}

\item{JsonSchema.variant}{The \link{JsonSchema.variant} object or list of objects}

\item{JsonSchema.variant.map}{The \link{JsonSchema.variant.map} object or list of objects}

\item{X.ref}{A reference to another schema}

\item{additionalProperties}{If this is a schema for an object, this property is the schema for any additional properties with dynamic keys on this object}

\item{annotations}{Additional information about this property}

\item{default}{The default value of this property (if one exists)}

\item{description}{A description of this object}

\item{enum}{Values this parameter may take (if it is an enum)}

\item{enumDescriptions}{The descriptions for the enums}

\item{format}{An additional regular expression or key that helps constrain the value}

\item{id}{Unique identifier for this schema}

\item{items}{If this is a schema for an array, this property is the schema for each element in the array}

\item{location}{Whether this parameter goes in the query or the path for REST requests}

\item{maximum}{The maximum value of this parameter}

\item{minimum}{The minimum value of this parameter}

\item{pattern}{The regular expression this parameter must conform to}

\item{properties}{If this is a schema for an object, list the schema for each property of this object}

\item{readOnly}{The value is read-only, generated by the service}

\item{repeated}{Whether this parameter may appear multiple times}

\item{required}{Whether the parameter is required}

\item{type}{The value type for this schema}

\item{variant}{In a variant data type, the value of one property is used to determine how to interpret the entire entity}
}
\value{
JsonSchema object
}
\description{
JsonSchema Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other JsonSchema functions: \code{\link{JsonSchema.annotations}},
  \code{\link{JsonSchema.properties}},
  \code{\link{JsonSchema.variant.map}},
  \code{\link{JsonSchema.variant}}
}
