% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery_objects.R
\name{JsonSchema.variant.map}
\alias{JsonSchema.variant.map}
\title{JsonSchema.variant.map Object}
\usage{
JsonSchema.variant.map()
}
\value{
JsonSchema.variant.map object
}
\description{
JsonSchema.variant.map Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other JsonSchema functions: \code{\link{JsonSchema.annotations}},
  \code{\link{JsonSchema.properties}},
  \code{\link{JsonSchema.variant}},
  \code{\link{JsonSchema}}
}
