% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery_objects.R
\name{JsonSchema.variant}
\alias{JsonSchema.variant}
\title{JsonSchema.variant Object}
\usage{
JsonSchema.variant(JsonSchema.variant.map = NULL, discriminant = NULL,
  map = NULL)
}
\arguments{
\item{JsonSchema.variant.map}{The \link{JsonSchema.variant.map} object or list of objects}

\item{discriminant}{The name of the type discriminant property}

\item{map}{The map of discriminant value to schema to use for parsing}
}
\value{
JsonSchema.variant object
}
\description{
JsonSchema.variant Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
In a variant data type, the value of one property is used to determine how to interpret the entire entity. Its value must exist in a map of descriminant values to schema names.
}
\seealso{
Other JsonSchema functions: \code{\link{JsonSchema.annotations}},
  \code{\link{JsonSchema.properties}},
  \code{\link{JsonSchema.variant.map}},
  \code{\link{JsonSchema}}
}
