% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery_objects.R
\name{RestDescription.auth.oauth2}
\alias{RestDescription.auth.oauth2}
\title{RestDescription.auth.oauth2 Object}
\usage{
RestDescription.auth.oauth2(RestDescription.auth.oauth2.scopes = NULL,
  scopes = NULL)
}
\arguments{
\item{RestDescription.auth.oauth2.scopes}{The \link{RestDescription.auth.oauth2.scopes} object or list of objects}

\item{scopes}{Available OAuth 2}
}
\value{
RestDescription.auth.oauth2 object
}
\description{
RestDescription.auth.oauth2 Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
OAuth 2.0 authentication information.
}
\seealso{
Other RestDescription functions: \code{\link{RestDescription.auth.oauth2.scopes}},
  \code{\link{RestDescription.auth}},
  \code{\link{RestDescription.icons}},
  \code{\link{RestDescription.methods}},
  \code{\link{RestDescription.parameters}},
  \code{\link{RestDescription.resources}},
  \code{\link{RestDescription.schemas}},
  \code{\link{RestDescription}}
}

