% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery_objects.R
\name{RestDescription.auth.oauth2.scopes}
\alias{RestDescription.auth.oauth2.scopes}
\title{RestDescription.auth.oauth2.scopes Object}
\usage{
RestDescription.auth.oauth2.scopes()
}
\value{
RestDescription.auth.oauth2.scopes object
}
\description{
RestDescription.auth.oauth2.scopes Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Available OAuth 2.0 scopes.
}
\seealso{
Other RestDescription functions: \code{\link{RestDescription.auth.oauth2}},
  \code{\link{RestDescription.auth}},
  \code{\link{RestDescription.icons}},
  \code{\link{RestDescription.methods}},
  \code{\link{RestDescription.parameters}},
  \code{\link{RestDescription.resources}},
  \code{\link{RestDescription.schemas}},
  \code{\link{RestDescription}}
}
