% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery_objects.R
\name{RestDescription.auth}
\alias{RestDescription.auth}
\title{RestDescription.auth Object}
\usage{
RestDescription.auth(RestDescription.auth.oauth2 = NULL,
  RestDescription.auth.oauth2.scopes = NULL, oauth2 = NULL)
}
\arguments{
\item{RestDescription.auth.oauth2}{The \link{RestDescription.auth.oauth2} object or list of objects}

\item{RestDescription.auth.oauth2.scopes}{The \link{RestDescription.auth.oauth2.scopes} object or list of objects}

\item{oauth2}{OAuth 2}
}
\value{
RestDescription.auth object
}
\description{
RestDescription.auth Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Authentication information.
}
\seealso{
Other RestDescription functions: \code{\link{RestDescription.auth.oauth2.scopes}},
  \code{\link{RestDescription.auth.oauth2}},
  \code{\link{RestDescription.icons}},
  \code{\link{RestDescription.methods}},
  \code{\link{RestDescription.parameters}},
  \code{\link{RestDescription.resources}},
  \code{\link{RestDescription.schemas}},
  \code{\link{RestDescription}}
}

