% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery_objects.R
\name{RestDescription.icons}
\alias{RestDescription.icons}
\title{RestDescription.icons Object}
\usage{
RestDescription.icons(x16 = NULL, x32 = NULL)
}
\arguments{
\item{x16}{The URL of the 16x16 icon}

\item{x32}{The URL of the 32x32 icon}
}
\value{
RestDescription.icons object
}
\description{
RestDescription.icons Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Links to 16x16 and 32x32 icons representing the API.
}
\seealso{
Other RestDescription functions: \code{\link{RestDescription.auth.oauth2.scopes}},
  \code{\link{RestDescription.auth.oauth2}},
  \code{\link{RestDescription.auth}},
  \code{\link{RestDescription.methods}},
  \code{\link{RestDescription.parameters}},
  \code{\link{RestDescription.resources}},
  \code{\link{RestDescription.schemas}},
  \code{\link{RestDescription}}
}

