% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery_objects.R
\name{RestDescription.methods}
\alias{RestDescription.methods}
\title{RestDescription.methods Object}
\usage{
RestDescription.methods()
}
\value{
RestDescription.methods object
}
\description{
RestDescription.methods Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
API-level methods for this API.
}
\seealso{
Other RestDescription functions: \code{\link{RestDescription.auth.oauth2.scopes}},
  \code{\link{RestDescription.auth.oauth2}},
  \code{\link{RestDescription.auth}},
  \code{\link{RestDescription.icons}},
  \code{\link{RestDescription.parameters}},
  \code{\link{RestDescription.resources}},
  \code{\link{RestDescription.schemas}},
  \code{\link{RestDescription}}
}

