% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery_objects.R
\name{RestDescription.parameters}
\alias{RestDescription.parameters}
\title{RestDescription.parameters Object}
\usage{
RestDescription.parameters()
}
\value{
RestDescription.parameters object
}
\description{
RestDescription.parameters Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Common parameters that apply across all apis.
}
\seealso{
Other RestDescription functions: \code{\link{RestDescription.auth.oauth2.scopes}},
  \code{\link{RestDescription.auth.oauth2}},
  \code{\link{RestDescription.auth}},
  \code{\link{RestDescription.icons}},
  \code{\link{RestDescription.methods}},
  \code{\link{RestDescription.resources}},
  \code{\link{RestDescription.schemas}},
  \code{\link{RestDescription}}
}

