% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery_objects.R
\name{RestDescription}
\alias{RestDescription}
\title{RestDescription Object}
\usage{
RestDescription(RestDescription.auth = NULL,
  RestDescription.auth.oauth2 = NULL,
  RestDescription.auth.oauth2.scopes = NULL, RestDescription.icons = NULL,
  RestDescription.methods = NULL, RestDescription.parameters = NULL,
  RestDescription.resources = NULL, RestDescription.schemas = NULL,
  auth = NULL, basePath = NULL, baseUrl = NULL, batchPath = NULL,
  canonicalName = NULL, description = NULL, discoveryVersion = NULL,
  documentationLink = NULL, exponentialBackoffDefault = NULL,
  features = NULL, icons = NULL, id = NULL, labels = NULL,
  methods = NULL, name = NULL, ownerDomain = NULL, ownerName = NULL,
  packagePath = NULL, parameters = NULL, protocol = NULL,
  resources = NULL, revision = NULL, rootUrl = NULL, schemas = NULL,
  servicePath = NULL, title = NULL, version = NULL,
  version_module = NULL)
}
\arguments{
\item{RestDescription.auth}{The \link{RestDescription.auth} object or list of objects}

\item{RestDescription.auth.oauth2}{The \link{RestDescription.auth.oauth2} object or list of objects}

\item{RestDescription.auth.oauth2.scopes}{The \link{RestDescription.auth.oauth2.scopes} object or list of objects}

\item{RestDescription.icons}{The \link{RestDescription.icons} object or list of objects}

\item{RestDescription.methods}{The \link{RestDescription.methods} object or list of objects}

\item{RestDescription.parameters}{The \link{RestDescription.parameters} object or list of objects}

\item{RestDescription.resources}{The \link{RestDescription.resources} object or list of objects}

\item{RestDescription.schemas}{The \link{RestDescription.schemas} object or list of objects}

\item{auth}{Authentication information}

\item{basePath}{[DEPRECATED] The base path for REST requests}

\item{baseUrl}{[DEPRECATED] The base URL for REST requests}

\item{batchPath}{The path for REST batch requests}

\item{canonicalName}{Indicates how the API name should be capitalized and split into various parts}

\item{description}{The description of this API}

\item{discoveryVersion}{Indicate the version of the Discovery API used to generate this doc}

\item{documentationLink}{A link to human readable documentation for the API}

\item{exponentialBackoffDefault}{Enable exponential backoff for suitable methods in the generated clients}

\item{features}{A list of supported features for this API}

\item{icons}{Links to 16x16 and 32x32 icons representing the API}

\item{id}{The ID of this API}

\item{labels}{Labels for the status of this API, such as labs or deprecated}

\item{methods}{API-level methods for this API}

\item{name}{The name of this API}

\item{ownerDomain}{The domain of the owner of this API}

\item{ownerName}{The name of the owner of this API}

\item{packagePath}{The package of the owner of this API}

\item{parameters}{Common parameters that apply across all apis}

\item{protocol}{The protocol described by this document}

\item{resources}{The resources in this API}

\item{revision}{The version of this API}

\item{rootUrl}{The root URL under which all API services live}

\item{schemas}{The schemas for this API}

\item{servicePath}{The base path for all REST requests}

\item{title}{The title of this API}

\item{version}{The version of this API}

\item{version_module}{No description}
}
\value{
RestDescription object
}
\description{
RestDescription Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other RestDescription functions: \code{\link{RestDescription.auth.oauth2.scopes}},
  \code{\link{RestDescription.auth.oauth2}},
  \code{\link{RestDescription.auth}},
  \code{\link{RestDescription.icons}},
  \code{\link{RestDescription.methods}},
  \code{\link{RestDescription.parameters}},
  \code{\link{RestDescription.resources}},
  \code{\link{RestDescription.schemas}}
}
