% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery_objects.R
\name{RestDescription.schemas}
\alias{RestDescription.schemas}
\title{RestDescription.schemas Object}
\usage{
RestDescription.schemas()
}
\value{
RestDescription.schemas object
}
\description{
RestDescription.schemas Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The schemas for this API.
}
\seealso{
Other RestDescription functions: \code{\link{RestDescription.auth.oauth2.scopes}},
  \code{\link{RestDescription.auth.oauth2}},
  \code{\link{RestDescription.auth}},
  \code{\link{RestDescription.icons}},
  \code{\link{RestDescription.methods}},
  \code{\link{RestDescription.parameters}},
  \code{\link{RestDescription.resources}},
  \code{\link{RestDescription}}
}

