% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery_objects.R
\name{RestMethod.mediaUpload.protocols}
\alias{RestMethod.mediaUpload.protocols}
\title{RestMethod.mediaUpload.protocols Object}
\usage{

  RestMethod.mediaUpload.protocols(RestMethod.mediaUpload.protocols.resumable = NULL,
  RestMethod.mediaUpload.protocols.simple = NULL, resumable = NULL,
  simple = NULL)
}
\arguments{
\item{RestMethod.mediaUpload.protocols.resumable}{The \link{RestMethod.mediaUpload.protocols.resumable} object or list of objects}

\item{RestMethod.mediaUpload.protocols.simple}{The \link{RestMethod.mediaUpload.protocols.simple} object or list of objects}

\item{resumable}{Supports the Resumable Media Upload protocol}

\item{simple}{Supports uploading as a single HTTP request}
}
\value{
RestMethod.mediaUpload.protocols object
}
\description{
RestMethod.mediaUpload.protocols Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Supported upload protocols.
}
\seealso{
Other RestMethod functions: \code{\link{RestMethod.mediaUpload.protocols.resumable}},
  \code{\link{RestMethod.mediaUpload.protocols.simple}},
  \code{\link{RestMethod.mediaUpload}},
  \code{\link{RestMethod.parameters}},
  \code{\link{RestMethod.request}},
  \code{\link{RestMethod.response}},
  \code{\link{RestMethod}}
}
