% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery_objects.R
\name{RestMethod.mediaUpload.protocols.resumable}
\alias{RestMethod.mediaUpload.protocols.resumable}
\title{RestMethod.mediaUpload.protocols.resumable Object}
\usage{
RestMethod.mediaUpload.protocols.resumable(multipart = NULL, path = NULL)
}
\arguments{
\item{multipart}{True if this endpoint supports uploading multipart media}

\item{path}{The URI path to be used for upload}
}
\value{
RestMethod.mediaUpload.protocols.resumable object
}
\description{
RestMethod.mediaUpload.protocols.resumable Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Supports the Resumable Media Upload protocol.
}
\seealso{
Other RestMethod functions: \code{\link{RestMethod.mediaUpload.protocols.simple}},
  \code{\link{RestMethod.mediaUpload.protocols}},
  \code{\link{RestMethod.mediaUpload}},
  \code{\link{RestMethod.parameters}},
  \code{\link{RestMethod.request}},
  \code{\link{RestMethod.response}},
  \code{\link{RestMethod}}
}

