% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery_objects.R
\name{RestMethod.mediaUpload.protocols.simple}
\alias{RestMethod.mediaUpload.protocols.simple}
\title{RestMethod.mediaUpload.protocols.simple Object}
\usage{
RestMethod.mediaUpload.protocols.simple(multipart = NULL, path = NULL)
}
\arguments{
\item{multipart}{True if this endpoint supports upload multipart media}

\item{path}{The URI path to be used for upload}
}
\value{
RestMethod.mediaUpload.protocols.simple object
}
\description{
RestMethod.mediaUpload.protocols.simple Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Supports uploading as a single HTTP request.
}
\seealso{
Other RestMethod functions: \code{\link{RestMethod.mediaUpload.protocols.resumable}},
  \code{\link{RestMethod.mediaUpload.protocols}},
  \code{\link{RestMethod.mediaUpload}},
  \code{\link{RestMethod.parameters}},
  \code{\link{RestMethod.request}},
  \code{\link{RestMethod.response}},
  \code{\link{RestMethod}}
}

