% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery_objects.R
\name{RestMethod.mediaUpload}
\alias{RestMethod.mediaUpload}
\title{RestMethod.mediaUpload Object}
\usage{
RestMethod.mediaUpload(RestMethod.mediaUpload.protocols = NULL,
  RestMethod.mediaUpload.protocols.resumable = NULL,
  RestMethod.mediaUpload.protocols.simple = NULL, accept = NULL,
  maxSize = NULL, protocols = NULL)
}
\arguments{
\item{RestMethod.mediaUpload.protocols}{The \link{RestMethod.mediaUpload.protocols} object or list of objects}

\item{RestMethod.mediaUpload.protocols.resumable}{The \link{RestMethod.mediaUpload.protocols.resumable} object or list of objects}

\item{RestMethod.mediaUpload.protocols.simple}{The \link{RestMethod.mediaUpload.protocols.simple} object or list of objects}

\item{accept}{MIME Media Ranges for acceptable media uploads to this method}

\item{maxSize}{Maximum size of a media upload, such as '1MB', '2GB' or '3TB'}

\item{protocols}{Supported upload protocols}
}
\value{
RestMethod.mediaUpload object
}
\description{
RestMethod.mediaUpload Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Media upload parameters.
}
\seealso{
Other RestMethod functions: \code{\link{RestMethod.mediaUpload.protocols.resumable}},
  \code{\link{RestMethod.mediaUpload.protocols.simple}},
  \code{\link{RestMethod.mediaUpload.protocols}},
  \code{\link{RestMethod.parameters}},
  \code{\link{RestMethod.request}},
  \code{\link{RestMethod.response}},
  \code{\link{RestMethod}}
}
