% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery_objects.R
\name{RestMethod.request}
\alias{RestMethod.request}
\title{RestMethod.request Object}
\usage{
RestMethod.request(X.ref = NULL, parameterName = NULL)
}
\arguments{
\item{X.ref}{Schema ID for the request schema}

\item{parameterName}{parameter name}
}
\value{
RestMethod.request object
}
\description{
RestMethod.request Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The schema for the request.
}
\seealso{
Other RestMethod functions: \code{\link{RestMethod.mediaUpload.protocols.resumable}},
  \code{\link{RestMethod.mediaUpload.protocols.simple}},
  \code{\link{RestMethod.mediaUpload.protocols}},
  \code{\link{RestMethod.mediaUpload}},
  \code{\link{RestMethod.parameters}},
  \code{\link{RestMethod.response}},
  \code{\link{RestMethod}}
}
