% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery_objects.R
\name{RestMethod.response}
\alias{RestMethod.response}
\title{RestMethod.response Object}
\usage{
RestMethod.response(X.ref = NULL)
}
\arguments{
\item{X.ref}{Schema ID for the response schema}
}
\value{
RestMethod.response object
}
\description{
RestMethod.response Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The schema for the response.
}
\seealso{
Other RestMethod functions: \code{\link{RestMethod.mediaUpload.protocols.resumable}},
  \code{\link{RestMethod.mediaUpload.protocols.simple}},
  \code{\link{RestMethod.mediaUpload.protocols}},
  \code{\link{RestMethod.mediaUpload}},
  \code{\link{RestMethod.parameters}},
  \code{\link{RestMethod.request}},
  \code{\link{RestMethod}}
}
