% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery_objects.R
\name{RestResource.methods}
\alias{RestResource.methods}
\title{RestResource.methods Object}
\usage{
RestResource.methods()
}
\value{
RestResource.methods object
}
\description{
RestResource.methods Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Methods on this resource.
}
\seealso{
Other RestResource functions: \code{\link{RestResource.resources}},
  \code{\link{RestResource}}
}
