% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery_objects.R
\name{RestResource}
\alias{RestResource}
\title{RestResource Object}
\usage{
RestResource(RestResource.methods = NULL, RestResource.resources = NULL,
  methods = NULL, resources = NULL)
}
\arguments{
\item{RestResource.methods}{The \link{RestResource.methods} object or list of objects}

\item{RestResource.resources}{The \link{RestResource.resources} object or list of objects}

\item{methods}{Methods on this resource}

\item{resources}{Sub-resources on this resource}
}
\value{
RestResource object
}
\description{
RestResource Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other RestResource functions: \code{\link{RestResource.methods}},
  \code{\link{RestResource.resources}}
}
