% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery_objects.R
\name{RestResource.resources}
\alias{RestResource.resources}
\title{RestResource.resources Object}
\usage{
RestResource.resources()
}
\value{
RestResource.resources object
}
\description{
RestResource.resources Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Sub-resources on this resource.
}
\seealso{
Other RestResource functions: \code{\link{RestResource.methods}},
  \code{\link{RestResource}}
}

